requirements.txt
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
cssselect==1.2.0
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
dnspython==2.7.0
exceptiongroup==1.2.2
fake-useragent==1.5.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
outcome==1.3.0.post0
packaging==24.1
parse==1.20.2
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.66.5
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.20
w3lib==2.2.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.20.2
